commit dc0cb1cda72a989d92d813e487cdff59f629aa3f
    6064 ixgbe needs X550 support
commit 2e1aefd1b4539a3d6cf03dbfa515eabf4de09f59
    7223 grub "setup" built-in broken by 6593

	Conflicts:
	exception_lists/hdrchk
	exception_lists/cstyle

+++ usr/src/grub/grub-0.97/stage2/fsys_reiserfs.c
@@ -1213,7 +1213,7 @@ reiserfs_dir (char *dirname)
 }
 
 int
-reiserfs_embed (int *start_sector, int needed_sectors)
+reiserfs_embed (unsigned long long *start_sector, int needed_sectors)
 {
   struct reiserfs_super_block super;
   int num_sectors;
